pyproject.toml
[project]
name = "ahead-take-home"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Leon" }]
dependencies = ["fastapi[standard]>=0.119.0", "sqlmodel>=0.0.27", "pydantic-settings>=2.11.0", "loguru>=0.7.3"]
requires-python = "==3.12.*"
readme = "README.md"
license = "MIT"


[project.optional-dependencies]
dev = ["ipykernel>=7.0", "pytest-asyncio>=1.2.0", "pytest-cov>=7.0.0"]


[tool.pdm]
distribution = false


[tool.pytest.ini_options]
log_level = "INFO"
asyncio_default_fixture_loop_scope = "function"
addopts = """
--exitfirst
--pdb
--pdbcls=IPython.terminal.debugger:TerminalPdb
-v
-r a
--cov
--cov-report=term
--cov-report=html
--cov-report=xml
--no-cov-on-fail
"""
